Code_Utsava/Code_Utsava/requirements.txt
# ---- Core Web Framework ----
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# ---- PDF & OCR Processing ----
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0

# ---- JSON & Utilities ----
regex==2024.11.6

# ---- Google Gemini API ----
google-genai==1.3.0

# ---- (Optional but Useful) ----
requests==2.32.3
numpy==2.1.2
